Move ByteSlicePool to oal.index
