+++ lucene/core/src/java/org/apache/lucene/index/ByteSliceReader.java
@@ -21,7 +21,6 @@ import org.apache.lucene.store.DataInput;
 import org.apache.lucene.store.DataOutput;
 import org.apache.lucene.util.BitUtil;
 import org.apache.lucene.util.ByteBlockPool;
-import org.apache.lucene.util.ByteSlicePool;
 
 /* IndexInput that knows how to read the byte slices written
  * by Posting and PostingVector.  We read the bytes in
