+++ lucene/core/src/test/org/apache/lucene/index/TestByteSliceReader.java
@@ -20,7 +20,6 @@ import java.util.Random;
 import org.apache.lucene.tests.util.LuceneTestCase;
 import org.apache.lucene.tests.util.TestUtil;
 import org.apache.lucene.util.ByteBlockPool;
-import org.apache.lucene.util.ByteSlicePool;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
